+++ src/helpers/utils.ts
@@ -1,7 +1,9 @@
-export type ArrayElements<TArray extends any[]> = TArray extends Array<infer TElement>
+export type ArrayElements<TArray extends ReadonlyArray<any>> = TArray extends ReadonlyArray<
+  infer TElement
+>
   ? TElement
   : never;
 
-export type UnionFromClasses<TClassesArray extends any[]> = InstanceType<
+export type UnionFromClasses<TClassesArray extends ReadonlyArray<any>> = InstanceType<
   ArrayElements<TClassesArray>
 >;